get rid of warnings

+++ Day11/Source.cpp
@@ -40,8 +40,8 @@ int main()
   //  cout << endl;
   //}
 
-  for (size_t i = 0; i < arr.size(); ++i) {
-    for (size_t j = 0; j < arr[i].size(); ++j) {
+  for (int i = 0; i < (int)arr.size(); ++i) {
+    for (int j = 0; j < (int)arr[i].size(); ++j) {
       if (isHourGlass(arr, i, j)) {
         max = std::max(max, sum(arr, i, j));
       }
